Add a LICENSE file
The SDK doesn't include a single LICENSE file, and instead includes licenses for the GoogleCastSDK's
third-party dependencies. In order for the iOS team's acknowledgements automation to work as expected,
the repository needs to have a single LICENSE file. This commit adds a script that combines the
third-party dependencies licenses into a single LICENSE file.